Unified all callbacks into one protocol. Sent some protocols and enums to a global header. Added UI Callbacks (still untested).
